+++ agrest-cayenne/src/test/java/io/agrest/cayenne/GET_ReadFilter_EntityEncoderFilterIT.java
@@ -186,7 +186,7 @@ public class GET_ReadFilter_EntityEncoderFilterIT extends DbTest {
                     .stage(SelectStage.APPLY_SERVER_PARAMS,
                             c -> RESOURCE_ENTITY_IS_FILTERED = c.getEntity().isFiltered())
                     .stage(SelectStage.ASSEMBLE_QUERY,
-                            c -> QUERY_PAGE_SIZE = CayenneProcessor.getQuery(c.getEntity()).getPageSize())
+                            c -> QUERY_PAGE_SIZE = CayenneProcessor.getCayenneEntity(c.getEntity()).getSelect().getPageSize())
                     .get();
         }