+++ RELEASE-NOTES.md
@@ -4,6 +4,7 @@
 * #493 Replace Constraint with per-entity API based on AgEntityOverlay
 * #497 Agrest fails to build on Java 11
 * #498 Made JAX-RS dependency "provided"
+* #500 Child relationship of a dynamic relationship fails to resolve
 * #501 Replace EntityEncoderFilter with per-entity object filter
 * #502 Per-entity CRUD filters and authorizers for UpdateBuilder
 * #503 New update stage - MAP_CHANGES